Combine ticker plugins, and remove Mt. Gox stuff since it's defunct now.
